Mast cells created along with secreted histamines. Game functionality modularized
