+++ client/components/GameFunctions/index.js
@@ -1,0 +1,4 @@
+import {preload, create} from './create'
+import {update} from './update'
+
+export {preload, create, update}
