+++ client/components/GameFunctions/preload.js
@@ -1,0 +1,7 @@
+// export function preload() {
+//     this.load.image('ship', 'assets/PNG/whitebloodcell.png')
+//     this.load.image('otherPlayer', 'assets/PNG/whitebloodcell.png')
+//     this.load.image('star', 'assets/PNG/star_gold.png')
+//     this.load.image('histamines', 'assets/PNG/Effects/star1.png')
+//     this.load.image('mastCell', 'assets/PNG/Meteors/meteorGrey_tiny2.png')
+//   }
