+++ client/components/phaser-game.js
@@ -1,8 +1,7 @@
 // var React = require('react-phaser')
 import Phaser from 'phaser'
 import React, {Component} from 'react'
-// un-comment the following line when cytofightFunctions.js is refactored
-// import {preload, create, update} from './cytofightFunctions'
+import {preload, create, update} from './GameFunctions/index'
 
 const config = {
   type: Phaser.AUTO,
@@ -22,150 +21,6 @@ const config = {
     preload: preload,
     create: create,
     update: update
-  }
-}
-
-function preload() {
-  this.load.image('ship', 'assets/PNG/whitebloodcell.png')
-  this.load.image('otherPlayer', 'assets/PNG/whitebloodcell.png')
-  this.load.image('star', 'assets/PNG/star_gold.png')
-}
-
-function create() {
-  const self = this
-  this.socket = io()
-  this.otherPlayers = this.physics.add.group()
-  this.socket.on('currentPlayers', function(players) {
-    Object.keys(players).forEach(function(id) {
-      if (players[id].playerId === self.socket.id) {
-        addPlayer(self, players[id])
-      } else {
-        addOtherPlayers(self, players[id])
-      }
-    })
-  })
-  this.socket.on('newPlayer', function(playerInfo) {
-    addOtherPlayers(self, playerInfo)
-  })
-  this.socket.on('disconnect', function(playerId) {
-    self.otherPlayers.getChildren().forEach(function(otherPlayer) {
-      if (playerId === otherPlayer.playerId) {
-        otherPlayer.destroy()
-      }
-    })
-  })
-  this.socket.on('playerMoved', function(playerInfo) {
-    self.otherPlayers.getChildren().forEach(function(otherPlayer) {
-      if (playerInfo.playerId === otherPlayer.playerId) {
-        otherPlayer.setRotation(playerInfo.rotation)
-        otherPlayer.setPosition(playerInfo.x, playerInfo.y)
-      }
-    })
-  })
-  this.cursors = this.input.keyboard.createCursorKeys()
-
-  //These 'WASD' keys need to be refractored so that the keycode numbers aren't hardcoded in. Phaser.KeyCode property doesn't exist?
-  this.keyLeft = this.input.keyboard.addKey(65);
-  this.keyRight = this.input.keyboard.addKey(68);
-  this.keyUp = this.input.keyboard.addKey(87);
-  this.keyDown = this.input.keyboard.addKey(83);
-  
-  this.blueScoreText = this.add.text(16, 16, '', { fontSize: '32px', fill: '#0000FF' });
-  this.redScoreText = this.add.text(584, 16, '', { fontSize: '32px', fill: '#FF0000' });
-  
-  this.socket.on('scoreUpdate', function (scores) {
-    self.blueScoreText.setText('Blue: ' + scores.blue);
-    self.redScoreText.setText('Red: ' + scores.red);
-  });
-
-  this.socket.on('starLocation', function (starLocation) {
-    if (self.star) self.star.destroy();
-    self.star = self.physics.add.image(starLocation.x, starLocation.y, 'star');
-    self.physics.add.overlap(self.ship, self.star, function () {
-      this.socket.emit('starCollected');
-    }, null, self);
-  });
-}
-
-function addPlayer(self, playerInfo) {
-  self.ship = self.physics.add
-    .image(playerInfo.x, playerInfo.y, 'ship')
-    .setOrigin(0.5, 0.5)
-    .setDisplaySize(53, 40)
-  if (playerInfo.team === 'blue') {
-    self.ship.setTint(0x0000ff)
-  } else {
-    self.ship.setTint(0xff0000)
-  }
-  self.ship.setDrag(100)
-  self.ship.setAngularDrag(100)
-  self.ship.setMaxVelocity(200)
-}
-
-function addOtherPlayers(self, playerInfo) {
-  const otherPlayer = self.add
-    .sprite(playerInfo.x, playerInfo.y, 'otherPlayer')
-    .setOrigin(0.5, 0.5)
-    .setDisplaySize(53, 40)
-  if (playerInfo.team === 'blue') {
-    otherPlayer.setTint(0x0000ff)
-  } else {
-    otherPlayer.setTint(0xff0000)
-  }
-  otherPlayer.playerId = playerInfo.playerId
-  self.otherPlayers.add(otherPlayer)
-}
-
-function update() {
-  if (this.ship) {
-    if (this.cursors.left.isDown || this.keyLeft.isDown) {
-      this.ship.setAngularVelocity(-150)
-    } else if (this.cursors.right.isDown || this.keyRight.isDown) {
-      this.ship.setAngularVelocity(150)
-    } else {
-      this.ship.setAngularVelocity(0)
-    }
-
-    if (this.cursors.up.isDown || this.keyUp.isDown) {
-      this.physics.velocityFromRotation(
-        this.ship.rotation + 1.5,
-        100,
-        this.ship.body.acceleration
-      )
-    } else if (this.cursors.down.isDown || this.keyDown.isDown) {
-      this.physics.velocityFromRotation(
-        this.ship.rotation + 1.5,
-        100,
-        this.ship.body.acceleration
-      )
-    } else {
-      this.ship.setAcceleration(0)
-    }
-    this.physics.world.wrap(this.ship, 5)
-    
-    // emit player movement
-    const x = this.ship.x
-    const y = this.ship.y
-    const r = this.ship.rotation
-    if (
-      this.ship.oldPosition &&
-      (x !== this.ship.oldPosition.x ||
-        y !== this.ship.oldPosition.y ||
-        r !== this.ship.oldPosition.rotation)
-    ) {
-      this.socket.emit('playerMovement', {
-        x: this.ship.x,
-        y: this.ship.y,
-        rotation: this.ship.rotation
-      })
-    }
-
-    // save old position data
-    this.ship.oldPosition = {
-      x: this.ship.x,
-      y: this.ship.y,
-      rotation: this.ship.rotation
-    }
   }
 }
